// assets/backend/admin/components/TimeSelector/Summary/Summary.module.css.d.ts
declare const styles:
  & Readonly<{ "summary": string }>
  & Readonly<{ "summary": string }>
  & Readonly<{ "summaryHeader": string }>
  & Readonly<{ "summaryBody": string }>
  & Readonly<{ "summaryRow": string }>
  & Readonly<{ "summaryRowError": string }>
;
export default styles;
//# sourceMappingURL=./Summary.module.css.d.ts.map